Add: funcations and test into seperate files, create fetch-calls.js
