+++ src/fetch-calls.js
@@ -1,0 +1,3 @@
+export const addRoomToBookings = () => { }
+
+export const deleteRoomFromBookings = () => { }
