--- test/sample-test.js
@@ -1,8 +1,0 @@
-import chai from 'chai';
-const expect = chai.expect;
-
-describe('See if the tests are running', function() {
-  it('should return true', function() {
-    expect(true).to.equal(true);
-  });
-});
